added photon-api service and injected it asynchronously in photo-location-field;
