+++ src/app/core/models/photon-query.model.ts
@@ -1,0 +1,5 @@
+export interface PhotonQuery {
+  q: string;
+  lang: 'en';
+  limit: number;
+}
